+++ src/readability.rs
@@ -4,9 +4,11 @@ use dom_query::{Document, Node, NodeData, NodeRef, Selection};
 use tendril::StrTendril;
 use url::Url;
 
+use crate::config::TextMode;
 use crate::glob::*;
 use crate::helpers::*;
 use crate::is_probably_readable;
+use crate::text_helpers::format_text;
 use crate::Config;
 use crate::ReadabilityError;
 
@@ -481,10 +483,17 @@ impl Readability {
             return Err(ReadabilityError::GrabFailed);
         };
 
+        let root_sel = doc.select_single_matcher(&MATCHER_CONTENT_ID);
+
+        let Some(root_node) = root_sel.nodes().first() else {
+            // After `grab_article` successfully returns a document
+            // there is no way that `#readability-page-1 does not exists
+            return Err(ReadabilityError::GrabFailed);
+        };
+
         // Getting a base uri from the Readability.document,
         // which wasn't changed after the grabbing the article
         let base_url = self.parse_base_url();
-        let root_sel = doc.select_single("#readability-page-1");
         self.post_process_content(&root_sel, base_url);
 
         // If we haven't found an excerpt in the article's metadata, use the article's
@@ -497,7 +506,10 @@ impl Readability {
             metadata.excerpt = extract_excerpt(&doc)
         }
 
-        let text_content = doc.text();
+        let text_content = match self.config.text_mode {
+            TextMode::Raw => root_node.text(),
+            TextMode::Formatted => format_text(root_node),
+        };
         let text_length = text_content.chars().count();
 
         Ok(Article {