+++ tests/alt.rs
@@ -1,0 +1,19 @@
+mod common;
+
+use std::fs;
+
+use common::test_alt_formatted_text;
+
+#[test]
+fn test_alt_formattedd_last_fail() {
+    test_alt_formatted_text("./test-pages/alt/hacker-news");
+}
+
+#[test]
+fn table_test_alt_formatted_text() {
+    let paths = fs::read_dir("./test-pages/alt").unwrap();
+    for p in paths {
+        let pp = p.unwrap().path();
+        test_alt_formatted_text(pp);
+    }
+}