+++ tests/metadata.rs
@@ -6,7 +6,10 @@ use common::test_metadata;
 
 #[test]
 fn test_metadata_last_fail() {
-    test_metadata("./test-pages/ok/arstechnica", Some("http://fakehost/test/"));
+    test_metadata(
+        "./test-pages/alt/arstechnica",
+        Some("http://fakehost/test/"),
+    );
 }
 
 #[test]
@@ -16,7 +19,6 @@ fn table_test_metadata() {
         let paths = fs::read_dir(d).unwrap();
         for p in paths {
             let pp = p.unwrap().path();
-            println!("path: {:?}", pp);
             test_metadata(pp, Some("http://fakehost/test/"));
         }
     }